backend/requirements.txt
openai
opencv-python
python-dotenv
SpeechRecognition
Flask
Flask-CORS
git+https://github.com/openai/whisper.git